move openCanvasDraw to CanvasDrawer component
